refactor: icons name
